Using size_t now for sizes

+++ libethash/ethash.h
@@ -41,8 +41,8 @@ extern "C" {
 
 typedef struct ethash_params
 {
-    unsigned full_size;					// Size of full data set (in bytes, multiple of page size (4096)).
-    unsigned cache_size;				// Size of compute cache (in bytes, multiple of node size (64)).
+    size_t full_size;					// Size of full data set (in bytes, multiple of page size (4096)).
+    size_t cache_size;				// Size of compute cache (in bytes, multiple of node size (64)).
 } ethash_params;
 
 uint32_t ethash_get_datasize(const uint32_t block_number);
